pyproject.toml
[project]
name = "test-giskard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "giskard[llm]>=2.18.0",
    "google-genai>=0.3.0",
    "python-dotenv>=1.2.1",
    "pandas>=2.0.0",
    "groq>=0.4.0",
    "fastembed<0.7.3",
    "pyarrow>=12.0.0,<15.0.0",
    "datasets>=2.14.0,<3.0.0",
    "ragas",
]
